+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+# 0.3.0 - 2016-03-??
+
+* Fix: counters cleared after performed action
+
 # 0.2.0 - 2016-03-16
 
 * Collect events with registered collector, syslog / null collectors supported